+++ example/lib/blob_sample.dart
@@ -87,14 +87,15 @@ class FallingBox extends BodyComponent {
 
 class BlobSample extends Forge2DGame with TapDetector {
   @override
-  bool debugMode() => true;
+  bool debugMode = true;
 
   BlobSample(Vector2 viewportSize)
       : super(
-          scale: 8.0,
+          scale: 20.0,
           gravity: Vector2(0, -10.0),
         ) {
     viewport.resize(viewportSize);
+    print("viewport: ${viewport.center}");
     final boundaries = createBoundaries(viewport);
     boundaries.forEach(add);
     add(Ground());
@@ -110,12 +111,25 @@ class BlobSample extends Forge2DGame with TapDetector {
     world.createJoint(jointDef);
   }
 
+  FallingBox box;
+  @override
+  void update(double dt) {
+    super.update(dt);
+    if(box != null) {
+      viewport.cameraFollow(box, horizontal: 1.0, vertical: 1.0);
+      print("box: ${box.body.position}");
+      print("world: ${viewport.getScreenToWorld(viewport.center)}");
+      print("viewport: ${viewport.center}");
+    }
+  }
+
   @override
   void onTapDown(TapDownDetails details) {
     super.onTapDown(details);
     final Vector2 screenPosition =
         Vector2(details.localPosition.dx, details.localPosition.dy);
     final Vector2 worldPosition = viewport.getScreenToWorld(screenPosition);
-    add(FallingBox(this, worldPosition));
+    box =FallingBox(this, worldPosition); 
+    add(box);
   }
 }