+++ example/lib/contact_callbacks_sample.dart
@@ -10,7 +10,7 @@ import 'boundaries.dart';
 
 class ContactCallbacksSample extends Forge2DGame with TapDetector {
   @override
-  bool debugMode() => true;
+  bool debugMode = true;
 
   ContactCallbacksSample(Vector2 viewportSize)
       : super(
@@ -23,13 +23,15 @@ class ContactCallbacksSample extends Forge2DGame with TapDetector {
     addContactCallback(BallContactCallback());
     addContactCallback(BallWallContactCallback());
     addContactCallback(WhiteBallContactCallback());
+    final Vector2 position = Vector2(100, 100);
+    add(Ball(position));
+    add(Ball(position));
   }
 
   @override
   void onTapDown(TapDownDetails details) {
     super.onTapDown(details);
-    final Vector2 position =
-        Vector2(details.localPosition.dx, details.localPosition.dy);
+    final Vector2 position = Vector2(100, 100);
     if (math.Random().nextInt(10) < 2) {
       add(WhiteBall(position));
     } else {