+++ example/lib/draggable_sample.dart
@@ -2,17 +2,17 @@ import 'package:flame_forge2d/flame_forge2d.dart';
 import 'package:flutter/gestures.dart';
 import 'package:forge2d/forge2d.dart';
 import 'package:flame_forge2d/forge2d_game.dart';
+import 'package:flame/components.dart';
+import 'package:flame/extensions.dart';
 import 'package:flame/palette.dart';
 import 'package:flame/gestures.dart';
-import 'package:flame/extensions/offset.dart';
-import 'package:flame/components/mixins/draggable.dart';
 
 import 'balls.dart';
 import 'boundaries.dart';
 
 class DraggableSample extends Forge2DGame with HasDraggableComponents {
   @override
-  bool debugMode() => true;
+  bool debugMode = true;
 
   DraggableSample(Vector2 viewportSize)
       : super(