+++ example/lib/sprite_body_sample.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
 import 'package:forge2d/forge2d.dart';
+import 'package:flame/components.dart';
 import 'package:flame/gestures.dart';
-import 'package:flame/sprite.dart';
 import 'package:flame_forge2d/forge2d_game.dart';
 import 'package:flame_forge2d/sprite_body_component.dart';
 import 'package:flutter/material.dart' hide Image;