+++ example/lib/tapable_sample.dart
@@ -4,14 +4,14 @@ import 'package:flutter/material.dart';
 import 'package:forge2d/forge2d.dart';
 import 'package:flame_forge2d/forge2d_game.dart';
 import 'package:flame/palette.dart';
-import 'package:flame/components/mixins/tapable.dart';
+import 'package:flame/components.dart';
 
 import 'balls.dart';
 import 'boundaries.dart';
 
 class TapableSample extends Forge2DGame with HasTapableComponents {
   @override
-  bool debugMode() => true;
+  bool debugMode = true;
 
   TapableSample(Vector2 viewportSize)
       : super(