+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
  - BodyComponent should follow Forge2D game debug mode
  - Fix generics for BodyComponent
+ - Add SpriteAnimationBodyComponent to display animated sprite with Forge2D
  
 ## [0.6.0-rc2]
  - Align with Flame 1.0.0-rc5